pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bucket_locker"
version = "0.1.0"
description = "A Python library that synchronizes local files with Google Cloud Storage in a concurrency-aware manner."
readme = "README.md"
authors = [
  {name = "Nadia Polikarpova", email = "nadia.polikarpova@gmail.com"}
]
classifiers = []
license = {text = "MIT license"}
dependencies = ["google-cloud-storage"]

[project.urls]
homepage = "https://github.com/nadia-polikarpova/bucket-locker"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

# ---- Dev & test deps ----
[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
]

# (Optional) make pytest auto-detect asyncio tests
[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"